LayoutTests/editing/undomanager/automatic-transaction-node.html

<!DOCTYPE html>
<html>
<head>
<title>automatic transaction on nodes</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
This tests that UndoManager correctly reverts/reapplies automatic DOM transactions
on node structural change.

<div id="edit"></div>
<script>
var edit = document.getElementById("edit");
var undoManager = document.undoManager;

test(function() {
    undoManager.transact({
        "executeAutomatic": function() {
            var d = document.createElement("span");
            d.innerText = "1";
            edit.appendChild(d);
        }
    });
    assert_true(edit.innerHTML == "<span>1</span>");
}, "UndoManager called the function returned by executeAutomatic. (appendChild)");

test(function() {
    undoManager.transact({
        "executeAutomatic": function() {
            var d2 = document.createElement("span");
            d2.innerHTML = "<b>2</b>";
            edit.insertBefore(d2, edit.firstChild);
        }
    });
    assert_true(edit.innerHTML == "<span><b>2</b></span><span>1</span>");
}, "UndoManager called the function returned by executeAutomatic. (insertBefore)");

test(function() {
    undoManager.transact({
        "executeAutomatic": function() {
            edit.removeChild(edit.lastChild);
        }
    });
    assert_true(edit.innerHTML == "<span><b>2</b></span>");
}, "UndoManager called the function returned by executeAutomatic. (removeChild)");

test(function() {
    undoManager.undo();
    assert_true(edit.innerHTML == "<span><b>2</b></span><span>1</span>");
}, "UndoManager reverted the change made by removeChild.");

test(function() {
    undoManager.undo();
    assert_true(edit.innerHTML == "<span>1</span>");
}, "UndoManager reverted the change made by insertBefore.");

test(function() {
    undoManager.undo();
    assert_true(edit.innerHTML == "");
}, "UndoManager reverted the change made by appendChild.");

test(function() {
    undoManager.redo();
    assert_true(edit.innerHTML == "<span>1</span>");
}, "UndoManager reapplied the change made by appendChild.");

test(function() {
    undoManager.redo();
    assert_true(edit.innerHTML == "<span><b>2</b></span><span>1</span>");
}, "UndoManager reapplied the change made by insertBefore.");

test(function() {
    undoManager.redo();
    assert_true(edit.innerHTML == "<span><b>2</b></span>");
}, "UndoManager reapplied the change made by removeChild.");

edit.style.display = "none";
</script>
</body>
</html>
